move subtasks into modules
