+++ src/24_game/solve.rs
