+++ src/averages/arithmetic_mean.rs
