+++ src/averages/root_mean_square.rs
