+++ src/closures/value_capture.rs
